pyproject.toml
[project]
name = "personal-job-market-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "playwright>=1.58.0",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
]
